커밋 v_0.0.08
1. 플레이어 HP 구현
2. 플레이어 체력 감소 코드 구현
3. 일정 시간 마다 플레이어의 체력이 닳는 코드 구현 및 적용
4. 저장 및 로드시 현재 플레이어의 체력이 같이 포함되도록 적용
5. 플레이어의 체력이 하나라도 0이하로 내려가면 게임 오버가 되도록 함
6..약간의 코드 간소화

+++ Assets/02_Scripts/DataManager.cs
@@ -7,9 +7,9 @@ using System.IO;
 public class PlayerData // 데이터를 저장할 클래스
 {
     public string UserName;
-    public int HungryHp;
-    public int WaterHp;
-    public int MentalHp;
+    public float HungryHp;
+    public float WaterHp;
+    public float MentalHp;
     public Vector3 playerPos;
     //public Vector3 playerRot; 
 }
